Fixed : operator not being allowed in symbols
Before, when you would try to use `https://github.com` as a symbol, for example,
it would not parse anything after the `:` as a part of the same symbol.

To fix this, I changed the `parse_symbol` function in `parse.rs` to accept any
text with `:` included. As a side effect, though, dictionaries/maps could not
be parsed.

```bash
{x: 0, y: 1}
```

To address this, I changed dictionaries to use `=` as a delimiter instead of `:`
so now dictionaries are written as follows:

```bash
{x=0, y=1}
```